requirements.txt
google_api_python_client==2.133.0
python-dotenv==1.0.1
Requests==2.32.3